Created a function to check for builtin commands and changed echo implementation to cover all edge cases
