+++ srcs/builtins/builtins.h
@@ -2,5 +2,6 @@
 #define BUILTINS_H
 
 int	echo(char **user_input);
+int	builtins_exec(char **argv);
 
 #endif
